frontend/requirements.txt
streamlit
requests
sqlalchemy
psycopg2-binary
python-dotenv
passlib
pandas
